# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Usuarios",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/api/usuarios/edit": {
      "put": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "actualizarUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/save": {
      "post": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "guardarUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "obtenerTodosUsuarios",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}": {
      "get": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "obtenerUsuarioPorID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/delete/{id}": {
      "delete": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "eliminarUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UsuarioDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombres": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "rut": {
            "type": "integer",
            "format": "int64"
          },
          "dv": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "correoElectronico": {
            "type": "string"
          },
          "contrasena": {
            "type": "string"
          }
        }
      }
    }
  }
}